+++ src/SourceString.h
@@ -16,7 +16,8 @@ public:
       cpp11::strings x,
       int skip = 0,
       bool skipEmptyRows = true,
-      const std::string& comment = "")
+      const std::string& comment = "",
+      bool skipQuotes = true)
       : string_(static_cast<SEXP>(x[0])) {
 
     begin_ = CHAR(string_);
@@ -26,7 +27,7 @@ public:
     begin_ = skipBom(begin_, end_);
 
     // Skip lines, if needed
-    begin_ = skipLines(begin_, end_, skip, skipEmptyRows, comment);
+    begin_ = skipLines(begin_, end_, skip, skipEmptyRows, comment, skipQuotes);
   }
 
   const char* begin() { return begin_; }